Fix -Werror=sign-compare errors in device code(#7164)
Not sure why these aren't being caught in local 10.2 envs or CI builds, but I can't build a local CUDA 11.0 env due to a mamba bug.

Authors:
  - ptaylor <paul.e.taylor@me.com>

Approvers:
  - Mark Harris (@harrism)
  - David (@davidwendt)

URL: https://github.com/rapidsai/cudf/pull/7164

+++ cpp/src/io/csv/datetime.cuh
@@ -130,7 +130,7 @@ __inline__ __device__ constexpr int32_t daysSinceBaseline(int year, int month, i
 __inline__ __device__ constexpr int32_t daysSinceEpoch(int year, int month, int day)
 {
   // Shift the start date to epoch to match unix time
-  static_assert(daysSinceBaseline(1970, 1, 1) == 719468_days,
+  static_assert(static_cast<uint32_t>(daysSinceBaseline(1970, 1, 1)) == 719468_days,
                 "Baseline to epoch returns incorrect number of days");
 
   return daysSinceBaseline(year, month, day) - daysSinceBaseline(1970, 1, 1);
